@@ -20,7 +20,8 @@ Bundler.require(*Rails.groups)
 module GetHelpToRetrain
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
-    config.load_defaults 5.2
+    config.load_defaults '6.0'
+    Rails.autoloaders.main.ignore("#{Rails.root}/app/admin")
 
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration can go into files in config/initializers
